requirements.txt
# ContextForge v2.0 Requirements
# Multi-Agent Research Gap Detector with FREE frameworks
# No API keys needed!

# Multi-Agent Frameworks
crewai>=0.28.0
crewai-tools>=0.1.6
langgraph>=1.0.2

# LangChain & Utilities
langchain>=1.0.5
langchain-community>=0.4.1
langchain-core>=0.1.0

# FREE LLM (Local via Ollama)
langchain-ollama>=0.1.0

# FREE Embeddings (HuggingFace)
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1

# Privacy & Security
diffprivlib>=0.6.6

# Document Processing
langchain-text-splitters>=0.0.1
pypdf2>=3.0.1
python-docx>=1.2.0

# Visualization & Dashboard
streamlit>=1.51.0
plotly>=5.17.0
altair>=5.0.0

# Data & Research
requests>=2.31.0
pandas>=2.3.3
numpy>=2.3.4
graphviz>=0.21

# Optional: GPU Support (uncomment if you have NVIDIA GPU)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0
